Move context related code to core.context package
