+++ mcmd/__main__.py
@@ -1,7 +1,7 @@
 import sys
 
-from mcmd.core.context_holder import set_context
-from mcmd.core.home_context import HomeContext
+from mcmd.core.context import set_context
+from mcmd.core.context.home_context import HomeContext
 
 MIN_PYTHON = (3, 6)
 if sys.version_info < MIN_PYTHON:
