+++ mcmd/commands/add.py
@@ -2,7 +2,7 @@ import mimetypes
 from os import path as os_path
 
 from mcmd.commands._registry import arguments
-from mcmd.core import context
+from mcmd.core.context import context
 from mcmd.core.command import command
 from mcmd.core.errors import McmdError
 from mcmd.io import io
