+++ mcmd/config/config.py
@@ -9,7 +9,6 @@ from pathlib import Path
 from ruamel.yaml import YAML
 
 import mcmd.core.errors as errors
-from mcmd.core import context
 
 _config = None
 _properties_file: Path = None
